Documentation: update link to image for newer Hugo
Recent versions of Hugo fails to render `troubleshooting.md` when the
`custom-metrics-elements.png` image isn't located in a static directory.

To avoid further complication with the `synchronize.sh` script from the
github.com/prometheus-operator/website project, images are now stored
under `Documentation/img` and the script will copy this directory to the
`static` directory of the website. Any content that wants to include an
image needs to follow the same approach.

The caveat is that the image will not be displayed from the GitHub UI. I
have tried to use page bundles [1] which would circumvent this
limitation but it would involve moving all the individual MarkDown files
into separate directories and rearrange the site's structure.

See also https://github.com/prometheus-operator/website/pull/17

[1] https://gohugo.io/content-management/page-bundles/

Signed-off-by: Simon Pasquier <spasquie@redhat.com>