Makefile: fix generate target
The 'generate' target needs to depend on the 'k8s-gen' target which
generates both type and client code. It allows the CI to verify that the
generated client is up-to-date.

Additionally I realized that the client-gen, lister-gen and informer-gen
commands have different behaviors depending on whether GOPATH is defined
and also whether the project is located inside GOPATH or not. In
practice, the GitHub actions CI didn't generate the files in
./pkg/client so it couldn't detect if the generated files drifted away
or not. The only way I've found to fix it is to generate the files in
the ./github.com/promehteus-operator/prometheus-operator/pkg/client/
directory and then manually move the content to the ./pkg/client
directory.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>