fix: invalid pod definition due to long volume names
Previously, adding secrets to Prometheus resource longer than 50
characters resulted in generation of statefulset that have
`volume.name` larger than 63 characters. This doesn't conform to
RFC-1123 (<= 63 chars) requirement and resulted in pod not getting
created.

This patch fixes the issue by ensuring that the volume name conform to
the requirement. Additionally it fixes an issue with SanitizeVolumeName
which used to truncate the name at DNS1123LabelMaxLength thus returing
same name for `<63-char>-foo` and `<63-char>-bar` to return unique name
by computing the sha1 and appending to the end.

This commit also ensures that secrets and configmaps provided via the
Alertmanager and Prometheus have correct volume mounts.

Bug: https://github.com/prometheus-operator/prometheus-operator/issues/4944

Signed-off-by: Simon Pasquier <spasquie@redhat.com>
Signed-off-by: Sunil Thaha <sthaha@redhat.com>