*: fix volume names for mounted secrets/configmaps
This change fixes a regression introduced in #4988 (84ffc50).

The volume names for additional secrets/configmaps provided by
`.spec.secrets` and `.spec.configMaps` shouldn't be suffixed by the hash
ID as it is done for the webconfig secrets/configmaps. Otherwise it
is impossible to mount them in a sidecar container since there's no way
to know the volume name in advance.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>