Main 59.1 sync (#5016)
* pkg/prometheus: add CAP_FOWNER capability to Thanos only if needed

When configuring the Thanos sidecar to upload blocks to object storage,
the container needs to the CAP_FOWNER capability. Otherwise it isn't
required. This is a small improvement on top of #4931 to avoid granting
more capabilities than needed.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>

* Fixed secret and configmap volume names that need to be mounted in additional containers (#5000)

* test/e2e: add testPromStrategicMergePatch

Signed-off-by: Simon Pasquier <spasquie@redhat.com>

* *: fix volume names for mounted secrets/configmaps

This change fixes a regression introduced in #4988 (84ffc50).

The volume names for additional secrets/configmaps provided by
`.spec.secrets` and `.spec.configMaps` shouldn't be suffixed by the hash
ID as it is done for the webconfig secrets/configmaps. Otherwise it
is impossible to mount them in a sidecar container since there's no way
to know the volume name in advance.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>

Signed-off-by: Simon Pasquier <spasquie@redhat.com>

* fix: thanos-sidecar CAP_FOWNER capability

this remove the CAP_ prefix of the FOWNER capability on thanos-sidecar

* 0.59.1 (#5013)

* docs: Update changelog for v0.59.1

* chore: Prep for v0.59.1

Signed-off-by: Simon Pasquier <spasquie@redhat.com>
Co-authored-by: Simon Pasquier <spasquie@redhat.com>
Co-authored-by: Francesco Steccanella <fsteccanella@gmail.com>