pkg/prometheus: remove FOWNER capability for Thanos sidecar (#5030)
After hearing from different users, it doesn't seem like the Thanos
sidecar needs the `FOWNER` capability. From a security standpoint, it's
better not to ask for a capability that isn't required. And even if it
is needed in some environments, it can be added using the strategic
merge patch technique.

Closes #5023

Signed-off-by: Simon Pasquier <spasquie@redhat.com>

Signed-off-by: Simon Pasquier <spasquie@redhat.com>