test: re-add e2e test for Prometheus Operator upgrade path (#5015)
* test: re-add e2e test for Prometheus Operator upgrade path

PR #4758 was reverted because it broke CI. CI failed because it
was out of sync with the main branch for a long time and some of
the function signature changed.

This commit is based on PR #4758 and fix CI function signature
error.

Fixes: #4067

Signed-off-by: heylongdacoder <heylongdacoder@gmail.com>

* test: handle Prometheus observedGeneration status check

Current version is 0.59.1. The Prometheus CRD observedGeneration
will be released at version 0.60.0. This commit set the framework
version to next minor version and only check observedGeneration
status when operator version >= 0.60.0.

Signed-off-by: heylongdacoder <heylongdacoder@gmail.com>

* test/framework: cleanup 0.57.0

Signed-off-by: heylongdacoder <heylongdacoder@gmail.com>

Signed-off-by: heylongdacoder <heylongdacoder@gmail.com>