pkg/prometheus: add `{grpc,http}ListenLocal` for Thanos sidecar
When `spec.thanos.listenLocal` is true, both gRPC and HTTP endpoints are
configured to listen on the loopback interface. It is sometimes needed
to listen on all interfaces for gRPC and on the loopback interface only
for HTTP.

This commit introduces the `spec.thanos.grpcListenLocal` and
`spec.thanos.httpListenLocal` fields to enable all possible use cases.
`spec.thanos.listenLocal` is marked as deprecated since the new fields
offer equivalent possibilities with fine-grained controls.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>