pkg/operator: avoid reconciliation for objects being deleted
Whenever the operator receives an update for a monitoring CR or a
statefulset that is marked for deletion, it is safer to wait for the
object being completely deleted before trying to reconcile.

Without this change, the operator may try to recreate a dependent
resource such as a configmap or secret owned by the being-deleted
resource which would create a race with the resource garbage collector.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>