Adds labels and annotations to PrometheusRule CR
Issue: https://github.com/prometheus-operator/prometheus-operator/issues/5050

Problem: described in the issue. Furthermore it also fixes the function
mergeKubectlAnnotations where previously it was working only because the
to map was being modified inside the function mergeMapsByPrefix.

Solution: Additions of labels and annotations to PrometheusRule CR that
are then merged into each rule. If a rule already defines a key it will
not be overwriten. Regarding mergeKubectlAnnotations function now
correctly sets result of mergingMaps to "to" instead of "from".
MergeMapsByPrefix now creates a new map so the orginals remain intact.

Signed-off-by: Joao Marcal <jmarcal@redhat.com>