pkg/promcfg: Add support for enableHttp2 flag in AlertingEndpoints
Among others "enable_http2" <bool> (default: true) should be configurable
for AlertingEndpoints in the prometheus configuration. At the moment
there is no option to overwrite the enable_http2 flag for the default prefix /
and it is always set to true by the prometheus-operator. If istio service-mesh
is in place, alerts will not be sent to alertmanager as long as enable_http2
is set to true (see issue https://github.com/prometheus/prometheus/issues/10213).
This PR adds the support to define the enableHttp2 flag in the AlertingEndpoint.

Fixes #5146

Signed-off-by: Raphael Husistein <raphael.husistein@hotmail.com>