Refactors version checks on prometheus/statefulset.go to use ConfigGenerator
Problem: depending on the Prometheus version used some commandline
arguments are valid where others aren't and the operator should log
a warning. Previously this was done with an "if/else" statement, however
since the operator now has ConfigGenerator we can use it to simplify the
code

Solution: start using the ConfigGenerator in prometheus/statefulset.
Unfortunately, when providing a default for a certain version or above it's not
entirely possible to get rid of the "if" part of the oginal "if/else"
statement

Signed-off-by: JoaoBraveCoding <jmarcal@redhat.com>