build: sign tagged images with cosign
Images with the `main` tag aren't signed because if I understand
correctly, it's not possible to sign again an existing image. Otherwise
we would have to suffix the tag with the short commit ID (e.g.
`:main-01234567`). At the same time, using the main (or latest) tag isn't
the best choice in terms of reproducibility/stability so not having
signed images is probably not a big deal.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>