fix: avoid panic when ScrapeConfig CRD isn't installed (#5550)
The operator would panic on a nil pointer exception if the ScrapeConfig
CRD isn't installed in the cluster. The root cause is the test in
`(*ResourceSelector).SelectScrapeConfigs()` that checked the interface
value against `nil`. The test would return `true` only if both the
interface's concrete type and value are nil. While the concrete value
was `nil`, the concrete type was `*informers.ForResource`.

Instead the controller will check that the `*informers.ForResource`
variable isn't nil before calling
`(*ResourceSelector).SelectScrapeConfigs()`.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>