chore: fix e2e tests
The new ScrapeConfig e2e tests (#5575) deployed the operator + admission
webhook + webhook configurations without cleaning the wehhook
configurations after the tests completed. The end result was that some
of the following tests were failing because the admission service wasn't
present anymore (e.g. the creation of PrometheusRules failed).

In practice, the admission webhook is not required for the ScrapeConfig
tests, hence this change deploys only the operator.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>