refactor: AddToStore functions do not shadow wrong usage
Previously if a nil store was passed to AddRemote*ToStore functions, it
would shadow the wrong usage and will only return an error which can't
be recovered. This patch removes that because we require the program to
crash/panic if the usage is wrong.

See: https://github.com/prometheus-operator/prometheus-operator/pull/5700#discussion_r1235027791

Signed-off-by: Sunil Thaha <sthaha@redhat.com>