doc: add hint why ServiceMonitor might not be in Service Discovery
Apparently ServiceMonitors will not be added to the Service Discovery
page if their selector does not match any Service. This is a bit
counter-intuitive, so let's just add this to the troubleshooting docs. I
found one other instance of this being confusing on Slack, which helped
me figure my problem out.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>