feat: optimize memory usage of secret/configmap informers
The secrets and configmaps informers only need to store the name,
namespace and resource version of the objects: the event handlers don't
look at the other metadata fields. By stripping the rest of the
metadata, we can save a fair share of memory on clusters with lots of
secrets/configmaps.

Experiments show that the maximum memory usage with/without the change
is similar but once the situation settles, we have a ~80% decrease in
memory usage with this commit (from 300MB -> 50MB). More details (with
graphs) in the pull request.

To take full advantage of this optimization, we also tune the GOGC
variable to "30" to reclaim the memory faster after startup.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>