alertmanager: add the continueStrategy option to manage alert routing
This feature adds an option called continueStrategy which defines how the continue in AlertManager will be managed. This includes a default option to always continue in the alerting routing tree which is the current strategy used. It also adds an option that doesn't change any value in continue config

Fixes #5781