feat: support `additionalArgs` in the Alertmanager CRD
Currently it is not possible to inject arbitrary arguments in an
Alertmanager running with the Operator. With this change, we unlock
arbitrary arguments that can be injected into the Alertmanager instance,
in the same way as is already supported for Prometheus and ThanosRuler
CRDs.

This will fully fix #4851 (which was partially solved in #4863).