docs: disableCompaction defaulting when using thanos sidecar
For the Prometheus CRD, the operator default to true for
`spec.disableCompaction` when `spec.thanos.objectStorageConfig(|File)`
is defined, but this is not documented anywhere but in code (and in
kube-prometheus-stack helm chart values, as it turns out).

(see the end of pkg/prometheus/server/statefulset.go).

Document the behavior in the API reference directly.

Signed-off-by: Max Gautier <mg@max.gautier.name>