Inline SafeTLSConfig CA, Cert, Key in scrape_config (#7359)
* Inline SafeTLSConfig CA,Cert,Key in scrape_config

The existing SafeTLSConfig specifies paths like:

```yaml
- job_name: scrapeConfig/default/testscrapeconfig1
  tls_config:
    ca_file: /etc/prometheus/certs/0_default_tls_ca
    cert_file: /etc/prometheus/certs/0_default_tls_cert
    key_file: /etc/prometheus/certs/0_default_tls_private-key
```

but these paths are not available to the scrapers.

This change will allow scrapers configured with http_sd (namely the OTel
Collector + Target Allocator setup) to be able to scrape endpoints that
require TLS certificates.

The new scrape configuration would look like:

```yaml
- job_name: scrapeConfig/default/testscrapeconfig1
  tls_config:
    ca: "ca-cert-info"
    cert: "cert-info"
    key: "<secret>"
```

Since the values are now inline in the scrape configuration, the
scrapers will be able to use it to scrape the endpoints with the
credentials it needs for secure communication.

- Updates test assets store to have secret information
- Updates golden files to recognize new values for tls_config

Fixes: https://github.com/open-telemetry/opentelemetry-operator/issues/3724

Signed-off-by: Charlie Le <charlie_le@apple.com>

* Add option in ConfigGenerator for inlining TLS Config

Signed-off-by: Charlie Le <charlie_le@apple.com>

---------

Signed-off-by: Charlie Le <charlie_le@apple.com>