docs: remote ingress monitoring guide
These docs were removed as the Ambassador project is not available
any longer and uses a license the prometheus-operator devs do not want
to promote.

Refer to https://kubernetes.github.io/ingress-nginx/user-guide/monitoring/#prometheus-and-grafana-installation-using-service-monitors
and https://prometheus-operator.dev/docs/platform/exposing-prometheus-and-alertmanager
for equivalent documentation that does not rely on Ambassador.