feat: optimize event handler
This commit improves the efficiency of event handlers on configuration
resources. In practice, the operator only triggers a reconciliation when
there's a change in the resource's generation or labels. It's a
pre-requisite to implement the status subresource for this type of
resources.

For secrets and configmaps, the operator still relies on the resource
version.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>