chore: update description of `minReadySeconds` fields
This commit also changes the field type from `*uint32` to `*int32` (with
a validation marker ensuring a greater than or equal to zero value) to
align with the Kubernetes API conventions. While theorically a breaking
change, it's safe to modify the type because nobody should be using
values larger than MaxInt32 (e.g. 2147483647).

Signed-off-by: Simon Pasquier <spasquie@redhat.com>