docs: Mention that portNumber behaves like a selector
The PodMonitor.spec.podMetricsEndpoints.portNumber field is confusing
because it does not actually tell Prometheus which port to scrape on a
target that has been matched by a pod selector.

Instead, it acts as a filter on
__meta_kubernetes_pod_container_port_number to select targets to scrape.

So if a Pod does not explicitly declare a Port with this number, the
target will be dropped by Prometheus, even though the Pod may accept
connections on that (undeclared) port.

This is particularly confusing because 'path', which is configured at
the same level, does tell Prometheus what to scrape and where.

Improve the docs to make this clearer.

With updates by Simon Pasquier

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>