feat: add new NoSelectedResources reason
This commit introduces the `NoSelectedResources` reason for the
`Reconciled` condition when a workload object selects no resources
(valid or invalid). It should make it easier to detect when resource
selectors are null or match no resource.

The `Alertmanager` resource is the exception because the configuration
may be provided only via a reference to a Secret key.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>